Unifying the JWE and JWK tests for unix and windows implementations.
