+++ UnitTests/jwk/JwkTest.cs
@@ -613,6 +613,22 @@ namespace UnitTests
             Assert.Equal(CngAlgorithm.ECDsaP256, test.Algorithm);
             Assert.True(test.IsEphemeral);
         }
+		
+		[Fact]
+        public void EccKey_Ecdh_Public()
+        {
+            //given
+            var key = new Jwk(crv: "P-256", x: "BHId3zoDv6pDgOUh8rKdloUZ0YumRTcaVDCppUPoYgk", y: "g3QIDhaWEksYtZ9OWjNHn9a6-i_P9o5_NrdISP0VWDU");
+
+            //when
+            var test = key.EcDiffieHellmanKey();
+
+            //then
+            Assert.NotNull(test);
+            Assert.Equal("P-256", key.Crv);
+            var curveName = test.ExportParameters(false).Curve.Oid.FriendlyName;
+            Assert.Equal("nistP256", curveName);
+        }
 
         [SkippableFact]
         public void EccKey_Cng_Private()
@@ -635,6 +651,26 @@ namespace UnitTests
             Assert.Equal(CngAlgorithm.ECDsaP256, test.Algorithm);
             Assert.True(test.IsEphemeral);
         }
+		
+        [Fact]
+        public void EccKey_Ecdh_Private()
+        {
+            //given
+            var key = new Jwk(crv: "P-256",
+                              x: "BHId3zoDv6pDgOUh8rKdloUZ0YumRTcaVDCppUPoYgk",
+                              y: "g3QIDhaWEksYtZ9OWjNHn9a6-i_P9o5_NrdISP0VWDU",
+                              d: "KpTnMOHEpskXvuXHFCfiRtGUHUZ9Dq5CCcZQ-19rYs4"
+                           );
+
+            //when
+            var test = key.EcDiffieHellmanKey();
+
+            //then
+            Assert.NotNull(test);
+            Assert.Equal("P-256", key.Crv);
+            var curveName = test.ExportParameters(false).Curve.Oid.FriendlyName;
+            Assert.Equal("nistP256", curveName);
+        }
 
         [SkippableFact]
         public void EccKey_Cng_Private_KeyAgreement()
@@ -656,6 +692,26 @@ namespace UnitTests
             Assert.Equal("P-256", key.Crv);
             Assert.Equal(CngAlgorithm.ECDiffieHellmanP256, test.Algorithm);
             Assert.True(test.IsEphemeral);
+        }
+		
+        [Fact]
+        public void EccKey_Ecdh_Private_KeyAgreement()
+        {
+            //given
+            var key = new Jwk(crv: "P-256",
+                              x: "BHId3zoDv6pDgOUh8rKdloUZ0YumRTcaVDCppUPoYgk",
+                              y: "g3QIDhaWEksYtZ9OWjNHn9a6-i_P9o5_NrdISP0VWDU",
+                              d: "KpTnMOHEpskXvuXHFCfiRtGUHUZ9Dq5CCcZQ-19rYs4"
+                           );
+
+            //when
+            var test = key.EcDiffieHellmanKey();
+
+            //then
+            Assert.NotNull(test);
+            Assert.Equal("P-256", key.Crv);
+            var curveName = test.ExportParameters(false).Curve.Oid.FriendlyName;
+            Assert.Equal("nistP256", curveName);
         }
 
         [Fact]
@@ -879,13 +935,18 @@ namespace UnitTests
             Assert.Null(test.D);
         }
 
-        [SkippableFact]
-        public void NewEccCng_Private_P256()
+        [SkippableTheory]
+        [InlineData("CNG")]
+        [InlineData("ECDH")]
+        public void NewEcc_Private_P256(string keyImplementation)
         {
-            Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            if (keyImplementation == "CNG")
+            {
+                Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            }
 
             //given
-            var test = new Jwk(Ecc256Private());
+            var test = keyImplementation == "CNG" ? new Jwk(Ecc256Private()) :  new Jwk(Ecc256PrivateEcdh());
 
             //then
             Assert.Equal(Jwk.KeyTypes.EC, test.Kty);
@@ -895,13 +956,18 @@ namespace UnitTests
             Assert.Equal("KpTnMOHEpskXvuXHFCfiRtGUHUZ9Dq5CCcZQ-19rYs4", test.D);
         }
 
-        [SkippableFact]
-        public void NewEccCng_Public_P384()
+        [SkippableTheory]
+        [InlineData("CNG")]
+        [InlineData("ECDH")]
+        public void NewEccCng_Public_P384(string keyImplementation)
         {
-            Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            if (keyImplementation == "CNG")
+            {
+                Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            }
 
             //given
-            var test = new Jwk(Ecc384Public(), false);
+            var test = keyImplementation == "CNG" ? new Jwk(Ecc384Public(), false) :  new Jwk(Ecc384PublicEcdh(), false);
 
             //then
             Assert.Equal(Jwk.KeyTypes.EC, test.Kty);
@@ -911,13 +977,18 @@ namespace UnitTests
             Assert.Null(test.D);
         }
 
-        [SkippableFact]
-        public void NewEccCng_Private_P384()
+        [SkippableTheory]
+        [InlineData("CNG")]
+        [InlineData("ECDH")]
+        public void NewEccCng_Private_P384(string keyImplementation)
         {
-            Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            if (keyImplementation == "CNG")
+            {
+                Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            }
 
             //given
-            var test = new Jwk(Ecc384Private());
+            var test = keyImplementation == "CNG" ? new Jwk(Ecc384Private()) : new Jwk(Ecc384PrivateEcdh());
 
             //then
             Assert.Equal(Jwk.KeyTypes.EC, test.Kty);
@@ -927,13 +998,18 @@ namespace UnitTests
             Assert.Equal("ice3abxagFJ0L6Fk3WHQQK33CSq6vbVuGOH-iEuc8tFe2joOIb4PUo3uz9afjPeL", test.D);
         }
 
-        [SkippableFact]
-        public void NewEccCng_Public_P521()
+        [SkippableTheory]
+        [InlineData("CNG")]
+        [InlineData("ECDH")]
+        public void NewEccCng_Public_P521(string keyImplementation)
         {
-            Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            if (keyImplementation == "CNG")
+            {
+                Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            }
 
             //given
-            var test = new Jwk(Ecc512Public(), false);
+            var test = keyImplementation == "CNG" ? new Jwk(Ecc512Public(), false) : new Jwk(Ecc512PublicEcdh(), false);
 
             //then
             Assert.Equal(Jwk.KeyTypes.EC, test.Kty);
@@ -943,13 +1019,18 @@ namespace UnitTests
             Assert.Null(test.D);
         }
 
-        [SkippableFact]
-        public void NewEccCng_Private_P521()
+        [SkippableTheory]
+        [InlineData("CNG")]
+        [InlineData("ECDH")]
+        public void NewEccCng_Private_P521(string keyImplementation)
         {
-            Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            if (keyImplementation == "CNG")
+            {
+                Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
+            }
 
             //given
-            var test = new Jwk(Ecc512Private());
+            var test = keyImplementation == "CNG" ? new Jwk(Ecc512Private()) : new Jwk(Ecc512PrivateEcdh());
 
             //then
             Assert.Equal(Jwk.KeyTypes.EC, test.Kty);
@@ -1002,7 +1083,7 @@ namespace UnitTests
         }
 
         [SkippableFact]
-        public void FromDictionary_EccKey()
+        public void FromDictionary_EccKey_Cng()
         {
             Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
 
@@ -1033,6 +1114,38 @@ namespace UnitTests
             Assert.NotNull(key);
             Assert.Equal(CngAlgorithm.ECDsaP256, key.Algorithm);
             Assert.True(key.IsEphemeral);
+        }
+		
+        [Fact]
+        public void FromDictionary_EccKey_Ecdh()
+        {
+            //given
+            var data = new Dictionary<string, object>
+            {
+                { "kty", "EC" },
+                { "use", "enc" },
+                { "crv", "P-256" },
+                { "x", "BHId3zoDv6pDgOUh8rKdloUZ0YumRTcaVDCppUPoYgk" },
+                { "y", "g3QIDhaWEksYtZ9OWjNHn9a6-i_P9o5_NrdISP0VWDU" },
+                { "d", "KpTnMOHEpskXvuXHFCfiRtGUHUZ9Dq5CCcZQ-19rYs4" }
+            };
+
+            //when
+            var test = Jwk.FromDictionary(data);
+
+            //then
+            Assert.Equal(Jwk.KeyTypes.EC, test.Kty);
+            Assert.Equal(Jwk.KeyUsage.Encryption, test.Use);
+            Assert.Equal("P-256", test.Crv);
+            Assert.Equal("BHId3zoDv6pDgOUh8rKdloUZ0YumRTcaVDCppUPoYgk", test.X);
+            Assert.Equal("g3QIDhaWEksYtZ9OWjNHn9a6-i_P9o5_NrdISP0VWDU", test.Y);
+            Assert.Equal("KpTnMOHEpskXvuXHFCfiRtGUHUZ9Dq5CCcZQ-19rYs4", test.D);
+
+            var key = test.EcDiffieHellmanKey();
+
+            Assert.NotNull(key);
+            var curveName = key.ExportParameters(false).Curve.Oid.FriendlyName;
+            Assert.Equal("nistP256", curveName);
         }
 
         [Fact]
@@ -1071,7 +1184,7 @@ namespace UnitTests
         }
 
         [SkippableFact]
-        public void FromJson_EccKey()
+        public void FromJson_EccKey_Cng()
         {
             Skip.IfNot(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "This requires CNG, which is Windows Only.");
 
@@ -1105,6 +1218,41 @@ namespace UnitTests
             Assert.NotNull(key);
             Assert.Equal(CngAlgorithm.ECDsaP256, key.Algorithm);
             Assert.True(key.IsEphemeral);
+        }
+		
+        [Fact]
+        public void FromJson_EccKey_Ecdh()
+        {
+            //given
+            var json = @"{
+                ""kty"": ""EC"",
+                ""kid"": ""Ex-p1KJFz8hQE1S76SzkhHcaObCKoDPrtAPJdWuTcTc"",
+                ""crv"": ""P-256"",
+                ""use"": ""enc"",
+                ""x"": ""BHId3zoDv6pDgOUh8rKdloUZ0YumRTcaVDCppUPoYgk"",
+                ""y"": ""g3QIDhaWEksYtZ9OWjNHn9a6-i_P9o5_NrdISP0VWDU"",
+                ""d"": ""KpTnMOHEpskXvuXHFCfiRtGUHUZ9Dq5CCcZQ-19rYs4""
+            }";
+
+            //when
+            var test = Jwk.FromJson(json, JWT.DefaultSettings.JsonMapper);
+
+            //then
+            Assert.Equal("Ex-p1KJFz8hQE1S76SzkhHcaObCKoDPrtAPJdWuTcTc", test.KeyId);
+            Assert.Equal(Jwk.KeyTypes.EC, test.Kty);
+            Assert.Equal(Jwk.KeyUsage.Encryption, test.Use);
+            Assert.Equal("P-256", test.Crv);
+            Assert.Equal("BHId3zoDv6pDgOUh8rKdloUZ0YumRTcaVDCppUPoYgk", test.X);
+            Assert.Equal("g3QIDhaWEksYtZ9OWjNHn9a6-i_P9o5_NrdISP0VWDU", test.Y);
+            Assert.Equal("KpTnMOHEpskXvuXHFCfiRtGUHUZ9Dq5CCcZQ-19rYs4", test.D);
+
+            var key = test.EcDiffieHellmanKey();
+
+            Assert.NotNull(key);
+
+            Assert.NotNull(key);
+            var curveName = key.ExportParameters(false).Curve.Oid.FriendlyName;
+            Assert.Equal("nistP256", curveName);
         }
 
         [Fact]
@@ -1254,14 +1402,30 @@ namespace UnitTests
 
             return EccKey.New(x, y, d, usage);
         }
-
-        private CngKey Ecc256Public(CngKeyUsages usage = CngKeyUsages.Signing)
+        
+        private ECDiffieHellman Ecc256PrivateEcdh()
         {
             byte[] x = { 4, 114, 29, 223, 58, 3, 191, 170, 67, 128, 229, 33, 242, 178, 157, 150, 133, 25, 209, 139, 166, 69, 55, 26, 84, 48, 169, 165, 67, 232, 98, 9 };
             byte[] y = { 131, 116, 8, 14, 22, 150, 18, 75, 24, 181, 159, 78, 90, 51, 71, 159, 214, 186, 250, 47, 207, 246, 142, 127, 54, 183, 72, 72, 253, 21, 88, 53 };
             byte[] d = { 42, 148, 231, 48, 225, 196, 166, 201, 23, 190, 229, 199, 20, 39, 226, 70, 209, 148, 29, 70, 125, 14, 174, 66, 9, 198, 80, 251, 95, 107, 98, 206 };
 
+            return EccKeyUnix.New(x, y, d);
+        }
+
+        private CngKey Ecc256Public(CngKeyUsages usage = CngKeyUsages.Signing)
+        {
+            byte[] x = { 4, 114, 29, 223, 58, 3, 191, 170, 67, 128, 229, 33, 242, 178, 157, 150, 133, 25, 209, 139, 166, 69, 55, 26, 84, 48, 169, 165, 67, 232, 98, 9 };
+            byte[] y = { 131, 116, 8, 14, 22, 150, 18, 75, 24, 181, 159, 78, 90, 51, 71, 159, 214, 186, 250, 47, 207, 246, 142, 127, 54, 183, 72, 72, 253, 21, 88, 53 };
+
             return EccKey.New(x, y, usage: usage);
+        }
+        
+        private ECDiffieHellman Ecc256PublicEcdh()
+        {
+            byte[] x = { 4, 114, 29, 223, 58, 3, 191, 170, 67, 128, 229, 33, 242, 178, 157, 150, 133, 25, 209, 139, 166, 69, 55, 26, 84, 48, 169, 165, 67, 232, 98, 9 };
+            byte[] y = { 131, 116, 8, 14, 22, 150, 18, 75, 24, 181, 159, 78, 90, 51, 71, 159, 214, 186, 250, 47, 207, 246, 142, 127, 54, 183, 72, 72, 253, 21, 88, 53 };
+
+            return EccKeyUnix.New(x, y);
         }
 
         private CngKey Ecc384Public()
@@ -1270,6 +1434,14 @@ namespace UnitTests
             byte[] y = { 189, 202, 196, 30, 153, 53, 22, 122, 171, 4, 188, 42, 71, 2, 9, 193, 191, 17, 111, 180, 78, 6, 110, 153, 240, 147, 203, 45, 152, 236, 181, 156, 232, 223, 227, 148, 68, 148, 221, 176, 57, 149, 44, 203, 83, 85, 75, 55 };
 
             return EccKey.New(x, y);
+        }
+        
+        private ECDiffieHellman Ecc384PublicEcdh()
+        {
+            byte[] x = { 70, 151, 220, 179, 62, 0, 79, 232, 114, 64, 58, 75, 91, 209, 232, 128, 7, 137, 151, 42, 13, 148, 15, 133, 93, 215, 7, 3, 136, 124, 14, 101, 242, 207, 192, 69, 212, 145, 88, 59, 222, 33, 127, 46, 30, 218, 175, 79 };
+            byte[] y = { 189, 202, 196, 30, 153, 53, 22, 122, 171, 4, 188, 42, 71, 2, 9, 193, 191, 17, 111, 180, 78, 6, 110, 153, 240, 147, 203, 45, 152, 236, 181, 156, 232, 223, 227, 148, 68, 148, 221, 176, 57, 149, 44, 203, 83, 85, 75, 55 };
+
+            return EccKeyUnix.New(x, y);
         }
 
         private CngKey Ecc384Private()
@@ -1280,6 +1452,15 @@ namespace UnitTests
 
             return EccKey.New(x, y, d);
         }
+        
+        private ECDiffieHellman Ecc384PrivateEcdh()
+        {
+            byte[] x = { 70, 151, 220, 179, 62, 0, 79, 232, 114, 64, 58, 75, 91, 209, 232, 128, 7, 137, 151, 42, 13, 148, 15, 133, 93, 215, 7, 3, 136, 124, 14, 101, 242, 207, 192, 69, 212, 145, 88, 59, 222, 33, 127, 46, 30, 218, 175, 79 };
+            byte[] y = { 189, 202, 196, 30, 153, 53, 22, 122, 171, 4, 188, 42, 71, 2, 9, 193, 191, 17, 111, 180, 78, 6, 110, 153, 240, 147, 203, 45, 152, 236, 181, 156, 232, 223, 227, 148, 68, 148, 221, 176, 57, 149, 44, 203, 83, 85, 75, 55 };
+            byte[] d = { 137, 199, 183, 105, 188, 90, 128, 82, 116, 47, 161, 100, 221, 97, 208, 64, 173, 247, 9, 42, 186, 189, 181, 110, 24, 225, 254, 136, 75, 156, 242, 209, 94, 218, 58, 14, 33, 190, 15, 82, 141, 238, 207, 214, 159, 140, 247, 139 };
+
+            return EccKeyUnix.New(x, y, d);
+        }
 
         private CngKey Ecc512Public()
         {
@@ -1287,6 +1468,14 @@ namespace UnitTests
             byte[] y = { 0, 60, 71, 97, 112, 106, 35, 121, 80, 182, 20, 167, 143, 8, 246, 108, 234, 160, 193, 10, 3, 148, 45, 11, 58, 177, 190, 172, 26, 178, 188, 240, 91, 25, 67, 79, 64, 241, 203, 65, 223, 218, 12, 227, 82, 178, 66, 160, 19, 194, 217, 172, 61, 250, 23, 78, 218, 130, 160, 105, 216, 208, 235, 124, 46, 32 };
 
             return EccKey.New(x, y);
+        }
+        
+        private ECDiffieHellman Ecc512PublicEcdh()
+        {
+            byte[] x = { 0, 248, 73, 203, 53, 184, 34, 69, 111, 217, 230, 255, 108, 212, 241, 229, 95, 239, 93, 131, 100, 37, 86, 152, 87, 98, 170, 43, 25, 35, 80, 137, 62, 112, 197, 113, 138, 116, 114, 55, 165, 128, 8, 139, 148, 237, 109, 121, 40, 205, 3, 61, 127, 28, 195, 58, 43, 228, 224, 228, 82, 224, 219, 148, 204, 96 };
+            byte[] y = { 0, 60, 71, 97, 112, 106, 35, 121, 80, 182, 20, 167, 143, 8, 246, 108, 234, 160, 193, 10, 3, 148, 45, 11, 58, 177, 190, 172, 26, 178, 188, 240, 91, 25, 67, 79, 64, 241, 203, 65, 223, 218, 12, 227, 82, 178, 66, 160, 19, 194, 217, 172, 61, 250, 23, 78, 218, 130, 160, 105, 216, 208, 235, 124, 46, 32 };
+
+            return EccKeyUnix.New(x, y);
         }
 
         private CngKey Ecc512Private()
@@ -1296,6 +1485,15 @@ namespace UnitTests
             byte[] d = { 0, 222, 129, 9, 133, 207, 123, 116, 176, 83, 95, 169, 29, 121, 160, 137, 22, 21, 176, 59, 203, 129, 62, 111, 19, 78, 14, 174, 20, 211, 56, 160, 83, 42, 74, 219, 208, 39, 231, 33, 84, 114, 71, 106, 109, 161, 116, 243, 166, 146, 252, 231, 137, 228, 99, 149, 152, 123, 201, 157, 155, 131, 181, 106, 179, 112 };
 
             return EccKey.New(x, y, d);
+        }
+        
+        private ECDiffieHellman Ecc512PrivateEcdh()
+        {
+            byte[] x = { 0, 248, 73, 203, 53, 184, 34, 69, 111, 217, 230, 255, 108, 212, 241, 229, 95, 239, 93, 131, 100, 37, 86, 152, 87, 98, 170, 43, 25, 35, 80, 137, 62, 112, 197, 113, 138, 116, 114, 55, 165, 128, 8, 139, 148, 237, 109, 121, 40, 205, 3, 61, 127, 28, 195, 58, 43, 228, 224, 228, 82, 224, 219, 148, 204, 96 };
+            byte[] y = { 0, 60, 71, 97, 112, 106, 35, 121, 80, 182, 20, 167, 143, 8, 246, 108, 234, 160, 193, 10, 3, 148, 45, 11, 58, 177, 190, 172, 26, 178, 188, 240, 91, 25, 67, 79, 64, 241, 203, 65, 223, 218, 12, 227, 82, 178, 66, 160, 19, 194, 217, 172, 61, 250, 23, 78, 218, 130, 160, 105, 216, 208, 235, 124, 46, 32 };
+            byte[] d = { 0, 222, 129, 9, 133, 207, 123, 116, 176, 83, 95, 169, 29, 121, 160, 137, 22, 21, 176, 59, 203, 129, 62, 111, 19, 78, 14, 174, 20, 211, 56, 160, 83, 42, 74, 219, 208, 39, 231, 33, 84, 114, 71, 106, 109, 161, 116, 243, 166, 146, 252, 231, 137, 228, 99, 149, 152, 123, 201, 157, 155, 131, 181, 106, 179, 112 };
+
+            return EccKeyUnix.New(x, y, d);
         }
 
         private ECDsa ECDSa256Public()
